src/app/core/validators/inscricao-estadual-estados/tocantins.validator.spec.ts
import { Tocantins } from './tocantins.validator';
import { InscricaoEstadualFactory } from './inscricao-estadual.factory';

describe('classe validacao: Inscricao Estadual Tocantins', () => {

  const estado = InscricaoEstadualFactory.criaInstancia(InscricaoEstadualFactory.TOCANTINS);

  it('instancia deve estar criada', () => {
    expect(estado).toBeTruthy();
  });

  it('inscricao estadual deve estar valida', () => {
    expect(estado.validar('56039877474')).toBe(true);
    expect(estado.validar('15035427652')).toBe(true);
    expect(estado.validar('17030340880')).toBe(true);
    expect(estado.validar('67018294950')).toBe(true);
  });

  it('inscricao estadual deve estar invalida', () => {
    expect(estado.validar('123461')).toBe(false);
    expect(estado.validar('31632')).toBe(false);
    expect(estado.validar('13461123462423461')).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for nulo', () => {
    expect(estado.validar(null)).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for undefined', () => {
    expect(estado.validar(undefined)).toBe(false);
  });

});
